feat(SimpleServer): add bare `gen_server`-like loop
This is more akin to the basic gen server implementation that is
described by Joe Armstrong in his book, where we have a TCO'd loop that
just receives messages and reacts to them. The main difference here is
that we know that we are handling a function coming in for `call` &
`cast`.

All in all this seems to be more consistent and saves us maybe 30-40 ms
for 200k insert messages in the `PriceServer`.